Enable "gcp" authentication structure for bearer token in .kube/config.
This adds the ability to parse the bearer token from ~/.kube/config
stored by Google Cloud Platform. This token is only good for an hour
since it was last renewed with a `kubectl` command. So if a 401 error
is returned adds a note to the exception message when the token is
expired.

Also updates other auth failure tests that were bypassing exception
parsing in Kubeclient::ClientMixing#handle_exception.