Fix @http_proxy_uri undefined warning
When not passing in a http_proxy_uri into the Kubeclient::Client.new,
calling most methods will then cause a flurry of warnings (when -w is
enabled) since it never is defined.

Setting it explicitly to nil if the http_proxy_uri argument is not
passed in will remove this warning.