Fix get_security_context_constraints warning
When the MissingKindCompatability class was introduced in
https://github.com/abonas/kubeclient/pull/214, it the mapping has a
plural form of the kind, instead of the singular form like most of the
other entries.  When this gets loaded in the
test/test_resource_list_without_kind.rb, it tries defining the
`get_security_context_constraints` twice:

* Once for the list action
* Once for the record

This means that you would not have both a singluar and plural method
name for the entities.method_names list, and they both would be plural.

To fix, do something similar to what was done for the `Endpoints` entry,
which is to have a rule for it in `Kubeclient::Common#parse_definition`