Tell Rubocop to target Ruby 2.2
When not configured, Rubocop normally defaults to 2.1
(oldest version Rubocop currently supports), but if
you have a `.ruby-version` file locally it uses that.

This caused confusing behavior where creating local
`.ruby-version` for rbenv/rvm/etc. in kubeclient dir
to test it under specific ruby also affects Rubocop,
while on Travis it was targetting 2.1 under all rubies.

Specifically, for 2.3 it wants `frozen_string_literal`
comment in all files, which we don't want (as long as
we support 2.2).