Enable refreshing of auth_provider bearer_token
Previously, once a Kubeclient::Context was created, auth_provider
bearer_tokens would be set and never refreshed.

This commit enables automatic refreshing of auth_provider bearer_tokens
by returning callables instead of static tokens. When used with the
faraday client, the callable is automatically called for each request,
and for the non-faraday client the callable is called before setting the
header manually.

Tests have been updated to cover this functionality for both client
types, and an additional test was added for refreshing
bearer_token_file because it did not appear to be covered before.

Co-authored-by: Jonathan Gnagy <jonathan@therubyist.org>
Co-authored-by: Hartley McGuire <skipkayhil@gmail.com>